Level 1 Capstone Task 1 - Nested Loops
This is my first Capstone Task from level 1 of the Python for Data Science.

This Capstone contains a task to which we nest a for loop in a pyramid format.

Comments have been included to describe how the code works. The file name of the Capstone task is Tritable. I've also included the methods of testing which helped in terms of finding a way to make this code work.